requirements.txt
Pillow>=9.0.0
opencv-python>=4.0.0
rich>=13.0.0
colorama>=0.4.0
numpy>=1.20.0
torch>=2.0.0 